--- a/packages/backend/package.json
+++ b/packages/backend/package.json
@@ -42,6 +42,7 @@
     "reflect-metadata": "^0.2.2",
     "tsyringe": "^4.10.0",
     "uuid": "^11.1.0",
+    "yaml": "^2.8.0",
     "zod": "^3.23.0"
   },
   "devDependencies": {