--- a/packages/frontend/src/pages/DashboardPage.css
+++ b/packages/frontend/src/pages/DashboardPage.css
@@ -1,13 +1,13 @@
+/* Dashboard Page Styles */
 .dashboard-page {
-  min-height: calc(100vh - 200px);
-  padding: 2rem 0;
+  min-height: 100vh;
   background-color: #f9fafb;
+  padding: 2rem 1rem;
 }
 
 .dashboard-container {
-  max-width: 1000px;
+  max-width: 1200px;
   margin: 0 auto;
-  padding: 0 1.5rem;
 }
 
 .dashboard-title {
@@ -15,6 +15,21 @@
   font-weight: 700;
   color: #111827;
   margin-bottom: 2rem;
+}
+
+/* Loading State */
+.loading-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  min-height: 60vh;
+  gap: 1rem;
+}
+
+.loading-container p {
+  color: #6b7280;
+  font-size: 1.125rem;
 }
 
 /* Dashboard Sections */
@@ -34,11 +49,12 @@
   background-color: white;
   border-radius: 0.75rem;
   padding: 1.5rem;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+  box-shadow: 0 1px 3px 0 rgba(0, 0, 0, 0.1);
 }
 
 .info-item {
   display: flex;
+  justify-content: space-between;
   align-items: center;
   padding: 0.75rem 0;
   border-bottom: 1px solid #e5e7eb;
@@ -51,22 +67,19 @@
 .info-label {
   font-weight: 500;
   color: #6b7280;
-  width: 150px;
 }
 
 .info-value {
+  font-weight: 600;
   color: #111827;
-  font-weight: 400;
 }
 
 .tier-badge {
-  display: inline-block;
-  padding: 0.25rem 0.75rem;
   background-color: #4f46e5;
   color: white;
-  border-radius: 0.375rem;
+  padding: 0.25rem 0.75rem;
+  border-radius: 9999px;
   font-size: 0.875rem;
-  font-weight: 500;
 }
 
 /* API Key Card */
@@ -74,7 +87,7 @@
   background-color: white;
   border-radius: 0.75rem;
   padding: 1.5rem;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+  box-shadow: 0 1px 3px 0 rgba(0, 0, 0, 0.1);
 }
 
 .api-key-display {
@@ -82,62 +95,53 @@
   align-items: center;
   gap: 0.5rem;
   margin-bottom: 1rem;
+  background-color: #f3f4f6;
+  padding: 0.75rem;
+  border-radius: 0.5rem;
 }
 
 .api-key-text {
   flex: 1;
-  padding: 0.75rem;
-  background-color: #f3f4f6;
-  border-radius: 0.375rem;
   font-family: monospace;
   font-size: 0.875rem;
-  overflow-x: auto;
-  white-space: nowrap;
+  color: #374151;
+  word-break: break-all;
 }
 
 .icon-button {
-  padding: 0.5rem;
-  background-color: transparent;
-  border: 1px solid #e5e7eb;
-  border-radius: 0.375rem;
-  cursor: pointer;
+  background: none;
+  border: none;
   font-size: 1.25rem;
-  transition: all 0.2s;
+  cursor: pointer;
+  padding: 0.25rem;
+  border-radius: 0.25rem;
+  transition: background-color 0.2s;
 }
 
 .icon-button:hover {
-  background-color: #f3f4f6;
-  border-color: #d1d5db;
+  background-color: #e5e7eb;
+}
+
+.icon-button:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
 }
 
 .success-message {
   color: #10b981;
   font-size: 0.875rem;
   margin-bottom: 0.5rem;
-  animation: fadeIn 0.3s ease-in;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-    transform: translateY(-10px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
 }
 
 .api-key-warning {
-  font-size: 0.875rem;
   color: #ef4444;
+  font-size: 0.875rem;
   margin: 0;
 }
 
 /* Usage Statistics */
 .usage-stats {
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
   gap: 1rem;
 }
 
@@ -145,7 +149,7 @@
   background-color: white;
   border-radius: 0.75rem;
   padding: 1.5rem;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+  box-shadow: 0 1px 3px 0 rgba(0, 0, 0, 0.1);
 }
 
 .usage-card h3 {
@@ -159,18 +163,19 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  margin-bottom: 0.5rem;
+  margin-bottom: 0.75rem;
   font-size: 0.875rem;
   color: #6b7280;
 }
 
 .usage-percentage {
   font-weight: 600;
+  color: #111827;
 }
 
 .usage-bar-container {
   width: 100%;
-  height: 8px;
+  height: 0.75rem;
   background-color: #e5e7eb;
   border-radius: 9999px;
   overflow: hidden;
@@ -179,8 +184,8 @@
 
 .usage-bar {
   height: 100%;
-  background-color: #10b981;
-  transition: width 0.3s ease-in-out;
+  background-color: #4f46e5;
+  transition: width 0.3s ease;
 }
 
 .usage-bar.near-limit {
@@ -197,27 +202,28 @@
 .dashboard-actions {
   display: flex;
   gap: 1rem;
-  justify-content: flex-end;
-  margin-top: 2rem;
+  flex-wrap: wrap;
 }
 
 .action-button {
   padding: 0.75rem 1.5rem;
-  border: none;
   border-radius: 0.5rem;
   font-weight: 600;
   font-size: 1rem;
+  border: none;
   cursor: pointer;
   transition: all 0.2s;
+  text-decoration: none;
+  display: inline-block;
 }
 
 .action-button.secondary {
-  background-color: #6b7280;
-  color: white;
+  background-color: #f3f4f6;
+  color: #374151;
 }
 
 .action-button.secondary:hover {
-  background-color: #4b5563;
+  background-color: #e5e7eb;
 }
 
 .action-button.danger {
@@ -229,96 +235,48 @@
   background-color: #dc2626;
 }
 
-/* Loading State */
-.loading-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  min-height: 400px;
-  gap: 1rem;
-}
-
-.spinner {
-  width: 40px;
-  height: 40px;
-  border: 3px solid #f3f3f3;
-  border-top: 3px solid #4f46e5;
-  border-radius: 50%;
-  animation: spin 1s linear infinite;
-}
-
-@keyframes spin {
-  0% { transform: rotate(0deg); }
-  100% { transform: rotate(360deg); }
-}
-
-/* Accessibility */
-.action-button:focus,
-.icon-button:focus {
-  outline: 3px solid #4f46e5;
-  outline-offset: 2px;
-}
-
-/* High contrast mode */
-@media (prefers-contrast: high) {
-  .info-card,
-  .api-key-card,
-  .usage-card {
-    border: 2px solid #111827;
-  }
-  
-  .icon-button {
-    border-width: 2px;
-  }
-}
-
-/* Reduced motion */
-@media (prefers-reduced-motion: reduce) {
-  .success-message,
-  .usage-bar,
-  .action-button,
-  .icon-button {
-    animation: none;
-    transition: none;
-  }
+.action-button:disabled {
+  opacity: 0.6;
+  cursor: not-allowed;
 }
 
 /* Responsive Design */
 @media (max-width: 768px) {
+  .dashboard-page {
+    padding: 1rem;
+  }
+
   .dashboard-title {
     font-size: 2rem;
   }
-  
+
+  .info-card,
+  .api-key-card,
+  .usage-card {
+    padding: 1rem;
+  }
+
   .info-item {
     flex-direction: column;
     align-items: flex-start;
     gap: 0.25rem;
   }
-  
-  .info-label {
-    width: auto;
-    font-size: 0.875rem;
-  }
-  
+
   .api-key-display {
     flex-wrap: wrap;
   }
-  
+
   .api-key-text {
     width: 100%;
-    font-size: 0.75rem;
+    margin-bottom: 0.5rem;
   }
-  
-  .usage-stats {
-    grid-template-columns: 1fr;
-  }
-  
+
   .dashboard-actions {
     flex-direction: column;
   }
-  
+
   .action-button {
     width: 100%;
+    text-align: center;
   }
 }