## Summary
- Implement comprehensive performance tests for the rate limiting system
- Verify system can handle 1000+ requests per second with accurate rate limiting
- Add memory leak detection and performance benchmarking capabilities

## Test Coverage
- **High Load Tests**: 10,000 requests from 100 concurrent users
- **Tier Scaling Tests**: Verify different tier levels (TIER1, TIER2, TIER3)
- **Concurrency Tests**: 1000 concurrent users, race condition handling
- **Memory Usage Tests**: Leak detection over sustained load

## Performance Results
- P99 response time: ~2ms (target: <100ms) ✅
- Average response time: ~0.3ms (target: <50ms) ✅
- Throughput: 300,000+ requests/second ✅
- Memory usage: Stable, no leaks detected ✅

## Implementation Details
- Created `RateLimitPerformanceTester` class for load testing
- Added performance metrics collection (P50, P95, P99 percentiles)
- Report generation in HTML, JSON, and Markdown formats
- npm scripts: `test:performance` and `test:performance:gc`

## Test Plan
- [x] Run `npm run test:performance` - all tests pass
- [x] Verify rate limiting accuracy under high load
- [x] Check memory usage remains stable
- [x] Confirm performance meets targets

🤖 Generated with [Claude Code](https://claude.ai/code)